chore: refactor out PlotSelector component
