+++ src/pages/PathwayDetailPage.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useCallback } from "react";
+import React, { useEffect, useState } from "react";
 import { useParams, Link } from "react-router-dom";
 import Markdown from "../components/Markdown";
 import { pathwayMetadata } from "../data/pathwayMetadata";
@@ -25,87 +25,13 @@ import {
   summarizeSummary,
 } from "../utils/timeseriesIndex";
 import PublicationBlock from "../components/PublicationBlock";
-import NormalizedStackedAreaChart from "../components/NormalizedStackedAreaChart";
-import MultiLineChart from "../components/MultiLineChart";
-import VerticalBarChart from "../components/VerticalBarChart";
-
-interface DataPoint {
-  sector: string;
-  metric: string;
-  year: string;
-  technology: string;
-  value: number;
-  unit: string;
-}
-
-interface TimeSeries {
-  data: DataPoint[];
-}
-
-type PlotType =
-  | "composition"
-  | "emissionsVolume"
-  | "emissionsEfficiency"
-  | "supply";
-
-const PLOT_OPTIONS = [
-  { value: "composition", label: "Energy Composition" },
-  { value: "emissionsVolume", label: "Emissions Volume" },
-  { value: "emissionsEfficiency", label: "Emissions Efficiency" },
-  { value: "supply", label: "Supply" },
-] as const;
+import { PlotSelector, TimeSeries } from "../components/PlotSelector";
 
 const PathwayDetailPage: React.FC = () => {
   const { id } = useParams<{ id: string }>();
   const [pathway, setPathway] = useState<PathwayMetadataType | null>(null);
   const [loading, setLoading] = useState(true);
-  const [selectedPlot, setSelectedPlot] = useState<PlotType>("composition");
   const [timeseriesdata, setTimeseriesdata] = useState<TimeSeries | null>(null);
-
-  // Helper function to check if data exists for a specific metric
-  const hasDataForMetric = useCallback(
-    (data: TimeSeries | null, metric: string): boolean => {
-      if (!data?.data) return false;
-      return data.data.some((d) => d.metric === metric);
-    },
-    [],
-  );
-
-  // Helper function to get available plot options based on data
-  const getAvailablePlotOptions = useCallback(
-    (data: TimeSeries | null): (typeof PLOT_OPTIONS)[number][] => {
-      if (!data) return [];
-
-      return PLOT_OPTIONS.filter((option) => {
-        switch (option.value) {
-          case "composition":
-            return hasDataForMetric(data, "capacity");
-          case "emissionsVolume":
-            return hasDataForMetric(data, "absoluteEmissions");
-          case "emissionsEfficiency":
-            return hasDataForMetric(data, "emissionsIntensity");
-          case "supply":
-            return hasDataForMetric(data, "capacity");
-          default:
-            return false;
-        }
-      });
-    },
-    [hasDataForMetric],
-  );
-
-  // Add effect to update selected plot if current selection becomes invalid
-  useEffect(() => {
-    if (timeseriesdata) {
-      const availableOptions = getAvailablePlotOptions(timeseriesdata);
-      if (
-        !availableOptions.find((opt) => opt.value === selectedPlot) &&
-        availableOptions.length > 0
-      ) {
-        setSelectedPlot(availableOptions[0].value);
-      }
-    }
-  }, [timeseriesdata, selectedPlot, getAvailablePlotOptions]);
 
   useEffect(() => {
     setLoading(true);
@@ -163,61 +89,6 @@ const PathwayDetailPage: React.FC = () => {
         .catch((error) => console.error("Error fetching JSON:", error));
     }
   }, [datasets]);
-
-  const renderPlot = () => {
-    if (!timeseriesdata || !datasets[0]) return null;
-
-    switch (selectedPlot) {
-      case "composition":
-        return (
-          <div className="flex flex-col items-center">
-            <NormalizedStackedAreaChart
-              key={`${datasets[0].datasetId}-composition`}
-              data={timeseriesdata}
-              width={450}
-              height={300}
-            />
-          </div>
-        );
-      case "emissionsVolume":
-        return (
-          <div className="flex flex-col items-center">
-            <VerticalBarChart
-              key={`${datasets[0].datasetId}-emissions-volume`}
-              data={timeseriesdata}
-              width={450}
-              height={300}
-              metric="absoluteEmissions"
-            />
-          </div>
-        );
-      case "emissionsEfficiency":
-        return (
-          <div className="flex flex-col items-center">
-            <VerticalBarChart
-              key={`${datasets[0].datasetId}-emissions-efficiency`}
-              data={timeseriesdata}
-              width={450}
-              height={300}
-              metric="emissionsIntensity"
-            />
-          </div>
-        );
-      case "supply":
-        return (
-          <div className="flex flex-col items-center">
-            <MultiLineChart
-              key={`${datasets[0].datasetId}-supply`}
-              data={timeseriesdata}
-              width={450}
-              height={300}
-            />
-          </div>
-        );
-      default:
-        return null;
-    }
-  };
 
   if (loading) {
     return (
@@ -346,37 +217,11 @@ const PathwayDetailPage: React.FC = () => {
             </div>
 
             <div className="md:col-span-5">
-              {timeseriesdata &&
-              getAvailablePlotOptions(timeseriesdata).length > 0 ? (
-                <div className="bg-neutral-50 border border-neutral-200 rounded-lg p-4 mb-6">
-                  <div className="flex flex-col mb-4">
-                    <label
-                      htmlFor="plot-select"
-                      className="text-sm font-medium text-rmigray-700 mb-2"
-                    >
-                      Select Plot
-                    </label>
-                    <select
-                      id="plot-select"
-                      value={selectedPlot}
-                      onChange={(e) =>
-                        setSelectedPlot(e.target.value as PlotType)
-                      }
-                      className="block w-full rounded-md border-rmigray-300 shadow-sm focus:border-energy focus:ring-energy sm:text-sm"
-                    >
-                      {getAvailablePlotOptions(timeseriesdata).map((option) => (
-                        <option
-                          key={option.value}
-                          value={option.value}
-                        >
-                          {option.label}
-                        </option>
-                      ))}
-                    </select>
-                  </div>
-                  <div className="mb-4">{renderPlot()}</div>
-                </div>
-              ) : null}
+              <PlotSelector
+                timeseriesdata={timeseriesdata}
+                datasetId={datasets[0]?.datasetId}
+                className="mb-6"
+              />
 
               <div className="bg-neutral-50 border border-neutral-200 rounded-lg p-4 mb-6">
                 <h3 className="text-lg font-medium text-rmigray-800 mb-3">
